Move cachingRequired to caching.js.
